+++ shadow/changelog.md
@@ -20,3 +20,4 @@ v 0.0.1
 - 添加不同的实现方式，并且可以基于点击进行切换
 - 添加切换table 和 view
 - 调整footer布局，如果没有sql执行，footer不展示
+- 抽取 VIEW_MODES 到factory ,从而支持footer toggle 可以进行动态的icon